Do not set workspace if it is `default`.
Terraform Cloud does not support workspaces. Running `terraform.py` with
a Terraform Cloud backend produces the error 'workspaces not supported'.

With this commit we invoke the command to set the namespace with the
`-no-color` option to receive a plain-text UTF-8 response without
terminal color codes. If the stderr response is empty or 'workspaces not
supported', we continue with program operation.