Do not set workspace env var not defined.
Terraform Cloud does not support workspaces when you specify a specific
workspace name. Running `terraform.py` with a Terraform Cloud backend
configured this way produces the error 'workspaces not supported'.

With this commit we only invoke the command to set the workspace if the
`TERRAFORM_WS_NAME` workspace environment variable is set.